src/requirements.txt
numpy
keras
tensorflow
flask-restful
requests
flickrapi
h5py
image
flask-cors
uwsgi
pillow